figures

+++ alisa/src/main/java/entity/figures/Triangle.java
@@ -4,6 +4,6 @@ package entity.figures;
  * Created by user on 04.04.2015.
  */
 public class Triangle extends Figure {
-    private int side;
+    private double side;
     private double height;
 }
